+++ node/clients/index.ts
@@ -1,0 +1,9 @@
+import { IOClients } from '@vtex/api'
+
+import SellerClient from './sellerClient'
+
+export class Clients extends IOClients {
+  public get sellerClient() {
+    return this.getOrSet('sellerClient', SellerClient)
+  }
+}