+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- node service to retrieve seller.
+
 ## [0.0.2] - 2023-07-10
 
 ## [0.0.1] - 2023-07-10
 
 ### Added
+
 - Initial release.